Move mapif functions from int_auction.c to mapif.c
